Renamed plot components
